#80 source code refactorings, moving business logic from controller to service
