+++ entity-web/src/main/java/eu/europeana/entity/web/controller/SearchController.java
@@ -1,13 +1,8 @@
 package eu.europeana.entity.web.controller;
 
-import java.util.ArrayList;
-import java.util.Collections;
-import java.util.List;
-
 import javax.annotation.Resource;
 import javax.servlet.http.HttpServletRequest;
 
-import org.apache.commons.lang3.ArrayUtils;
 import org.apache.commons.lang3.StringUtils;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -25,18 +20,15 @@ import eu.europeana.api.commons.config.i18n.I18nConstants;
 import eu.europeana.api.commons.definitions.search.Query;
 import eu.europeana.api.commons.definitions.search.ResultSet;
 import eu.europeana.api.commons.definitions.search.result.ResultsPage;
-import eu.europeana.api.commons.definitions.search.result.impl.ResultsPageImpl;
 import eu.europeana.api.commons.definitions.vocabulary.CommonApiConstants;
 import eu.europeana.api.commons.definitions.vocabulary.CommonLdConstants;
 import eu.europeana.api.commons.definitions.vocabulary.ContextTypes;
-import eu.europeana.api.commons.search.util.QueryBuilder;
 import eu.europeana.api.commons.utils.ResultsPageSerializer;
 import eu.europeana.api.commons.web.exception.HttpException;
 import eu.europeana.api.commons.web.http.HttpHeaders;
 import eu.europeana.entity.definitions.exceptions.UnsupportedEntityTypeException;
 import eu.europeana.entity.definitions.model.Entity;
 import eu.europeana.entity.definitions.model.search.SearchProfiles;
-import eu.europeana.entity.definitions.model.vocabulary.ConceptSolrFields;
 import eu.europeana.entity.definitions.model.vocabulary.EntityTypes;
 import eu.europeana.entity.definitions.model.vocabulary.WebEntityConstants;
 import eu.europeana.entity.web.exception.InternalServerException;
@@ -161,10 +153,10 @@ public class SearchController extends BaseRest {
 			} 
 			
 			// perform search
-			Query searchQuery = buildSearchQuery(queryString, qf, facets, sort, page, pageSize, searchProfile, retFields);
-
+			Query searchQuery = entityService.buildSearchQuery(queryString, qf, facets, sort, page, pageSize, searchProfile, retFields);
 			ResultSet<? extends Entity> results = entityService.search(searchQuery, preferredLanguages, entityTypes, scope);
-			ResultsPage<? extends Entity> resPage = buildResultsPage(searchQuery, results, request.getRequestURL(),
+
+			ResultsPage<? extends Entity> resPage = entityService.buildResultsPage(searchQuery, results, request.getRequestURL(),
 					request.getQueryString());
 			String jsonLd = searializeResultsPage(resPage, searchProfile);
 
@@ -193,200 +185,6 @@ public class SearchController extends BaseRest {
 			throw new InternalServerException(e);
 		}
 	}
-
-	private String validateScopeParam(String scope) throws ParamValidationException {
-		if (StringUtils.isBlank(scope))
-			return null;
-				
-		if (!WebEntityConstants.PARAM_SCOPE_EUROPEANA.equalsIgnoreCase(scope))
-			throw new ParamValidationException(I18nConstants.INVALID_PARAM_VALUE,
-					WebEntityConstants.QUERY_PARAM_SCOPE, scope);
-		
-		return WebEntityConstants.PARAM_SCOPE_EUROPEANA;
-	}
-
-	private String searializeResultsPage(ResultsPage<? extends Entity> resPage, SearchProfiles profile)
-			throws JsonProcessingException {
-		ResultsPageSerializer<? extends Entity> serializer = new EntityResultsPageSerializer<>(resPage,
-				ContextTypes.ENTITY.getJsonValue(), CommonLdConstants.COLLECTION);
-		String profileVal = (profile == null)? null : profile.name(); 
-		return serializer.serialize(profileVal);
-	}
-
-	// TODO move to helper class
-	private <T extends Entity> ResultsPage<T> buildResultsPage(Query searchQuery, ResultSet<T> results,
-			StringBuffer requestUrl, String reqParams) {
-		@SuppressWarnings({ "rawtypes", "unchecked" })
-		ResultsPage<T> resPage = new ResultsPageImpl();
-
-		resPage.setItems(results.getResults());
-		resPage.setFacetFields(results.getFacetFields());
-
-		resPage.setTotalInPage(results.getResults().size());
-		resPage.setTotalInCollection(results.getResultSize());
-
-		String collectionUrl = buildCollectionUrl(searchQuery, requestUrl, reqParams);
-		resPage.setCollectionUri(collectionUrl);
-
-		int currentPage = searchQuery.getPageNr();
-		String currentPageUrl = buildPageUrl(collectionUrl, currentPage, searchQuery.getPageSize());
-		resPage.setCurrentPageUri(currentPageUrl);
-
-		if (currentPage > 0) {
-			String prevPage = buildPageUrl(collectionUrl, currentPage - 1, searchQuery.getPageSize());
-			resPage.setPrevPageUri(prevPage);
-		}
-
-		// if current page is not the last one
-		boolean isLastPage = resPage.getTotalInCollection() <= (currentPage + 1) * searchQuery.getPageSize();
-		if (!isLastPage) {
-			String nextPage = buildPageUrl(collectionUrl, currentPage + 1, searchQuery.getPageSize());
-			resPage.setNextPageUri(nextPage);
-		}
-
-		return resPage;
-	}
-
-	private String buildPageUrl(String collectionUrl, int page, int pageSize) {
-		StringBuilder builder = new StringBuilder(collectionUrl);
-		builder.append("&").append(CommonApiConstants.QUERY_PARAM_PAGE).append("=").append(page);
-
-		builder.append("&").append(CommonApiConstants.QUERY_PARAM_PAGE_SIZE).append("=").append(pageSize);
-
-		return builder.toString();
-	}
-
-	private String buildCollectionUrl(Query searchQuery, StringBuffer requestUrl, String queryString) {
-
-		// queryString = removeParam(WebAnnotationFields.PARAM_WSKEY,
-		// queryString);
-
-		// remove out of scope parameters
-		queryString = removeParam(CommonApiConstants.QUERY_PARAM_PAGE, queryString);
-		queryString = removeParam(CommonApiConstants.QUERY_PARAM_PAGE_SIZE, queryString);
-
-		// avoid duplication of query parameters
-		queryString = removeParam(CommonApiConstants.QUERY_PARAM_PROFILE, queryString);
-
-		// add mandatory parameters
-		if (StringUtils.isNotBlank(searchQuery.getSearchProfile())) {
-			queryString += ("&" + CommonApiConstants.QUERY_PARAM_PROFILE + "=" + searchQuery.getSearchProfile());
-		}
-
-		return requestUrl.append("?").append(queryString).toString();
-	}
-
-	protected String removeParam(final String queryParam, String queryParams) {
-		String tmp;
-		// avoid name conflicts search "queryParam="
-		int startPos = queryParams.indexOf(queryParam + "=");
-		int startEndPos = queryParams.indexOf("&", startPos + 1);
-
-		if (startPos >= 0) {
-			// make sure to remove the "&" if not the first param
-			if (startPos > 0)
-				startPos--;
-			tmp = queryParams.substring(0, startPos);
-
-			if (startEndPos > 0)
-				tmp += queryParams.substring(startEndPos);
-		} else {
-			tmp = queryParams;
-		}
-		return tmp;
-	}
-
-	/**
-	 * This method enriches provided custom selection fields by required fields
-	 * if they are not already provided in input array.
-	 * @param inputArray
-	 * @return enriched array
-	 */
-	private String[] buildCustomSelectionFields(String[] inputFields)  
-	{  
-		List<String> fieldList = new ArrayList<String>();
-		String[] fixedFields = {ConceptSolrFields.ID, ConceptSolrFields.INTERNAL_TYPE};
-
-		Collections.addAll(fieldList, inputFields);
-		Collections.addAll(fieldList, fixedFields);
-		
-		String [] resultFields = new String[fieldList.size()];
-		fieldList.toArray(resultFields);
-		
-		return resultFields;
-	}  
-
-	/**
-	 * @param queryString
-	 * @param qf
-	 * @param facets
-	 * @param sort
-	 * @param page
-	 * @param pageSize
-	 * @param profile
-	 * @param retFields
-	 * @return
-	 */
-	private Query buildSearchQuery(String queryString, String[] qf, String[] facets, String sort, int page,
-			int pageSize, SearchProfiles profile, String[] retFields) {
-		String sortField = null;
-		String sortOrder = null;
-		if (StringUtils.isNotBlank(sort)) {
-			String[] sorting = StringUtils.split(sort, '+');
-			sortField = sorting[0];
-			if (sorting.length > 1)
-				sortOrder = sorting[1];
-		}
-
-		QueryBuilder builder = new QueryBuilder();
-		int maxPageSize = Query.DEFAULT_MAX_PAGE_SIZE;
-		String profileName = null;
-		if (profile != null) {
-			profileName = profile.name();
-		}
-		
-		if (retFields != null) {
-			retFields = buildCustomSelectionFields(retFields);
-		}
-
-		Query query = builder.buildSearchQuery(queryString, qf, facets, retFields, sortField, sortOrder, page, pageSize,
-				maxPageSize, profileName);
-		return query;
-	}
-
-	/**
-	 * Get entity type string list from comma separated entities string.
-	 * 
-	 * @param commaSepEntityTypes
-	 *            Comma separated entities string
-	 * @return Entity types string list
-	 * @throws ParamValidationException
-	 */
-	public EntityTypes[] getEntityTypesFromString(String commaSepEntityTypes) throws ParamValidationException {
-
-		String[] splittedEntityTypes = commaSepEntityTypes.split(",");
-		EntityTypes[] entityTypes = new EntityTypes[splittedEntityTypes.length];
-
-		EntityTypes entityType = null;
-		String typeAsString = null;
-
-		try {
-			for (int i = 0; i < splittedEntityTypes.length; i++) {
-				typeAsString = splittedEntityTypes[i].trim();
-				entityType = EntityTypes.getByInternalType(typeAsString);
-				entityTypes[i] = entityType;
-			}
-		} catch (UnsupportedEntityTypeException e) {
-			throw new ParamValidationException(I18nConstants.INVALID_PARAM_VALUE, WebEntityConstants.QUERY_PARAM_TYPE,
-					typeAsString);
-		}
-
-		return entityTypes;
-	}
-
-	String[] getLanguageList(String requestedLanguages){
-		String[] languageArray = StringUtils.splitByWholeSeparator(requestedLanguages, ",");
-		return StringUtils.stripAll(languageArray);	
-	}
+	
 	
 }
